Navigation Module Complete 90% Feature Module Complete 80% (Created cart and order list) App Module Complete 80%

+++ features/orders/src/main/java/com/example/orders/domain/entities/OrderItem.kt
@@ -3,6 +3,6 @@ package com.example.orders.domain.entities
 data class OrderItem(
     val id: String,
     val name: String,
-    val quantity: Double,
+    val quantity: Int,
     val price: Price,
 )
